Fetch last messages from the user table
Fetching all messages only to return the last is really slow, this commit changes that so the last message should be stored in the user table.

Also add a decorator function to make a function "fire and forget". If the result of the function is not important for the return value, you can just fire the function and continue right away without waiting